+++ include/metalang99.h
@@ -14,6 +14,7 @@
 #include <metalang99/list.h>
 #include <metalang99/maybe.h>
 #include <metalang99/misc.h>
+#include <metalang99/tuple.h>
 #include <metalang99/uint.h>
 #include <metalang99/util.h>
 #include <metalang99/variadics.h>